+++ server/src/test/java/org/opensearch/cluster/metadata/MetadataCreateIndexServiceTests.java
@@ -957,13 +957,7 @@ public class MetadataCreateIndexServiceTests extends OpenSearchTestCase {
         assertThat(
             expectThrows(
                 IllegalStateException.class,
-                () -> clusterStateCreateIndex(
-                    currentClusterState,
-                    org.opensearch.common.collect.Set.of(),
-                    newIndex,
-                    (state, reason) -> state,
-                    null
-                )
+                () -> clusterStateCreateIndex(currentClusterState, Set.of(), newIndex, (state, reason) -> state, null)
             ).getMessage(),
             startsWith("alias [alias1] has more than one write index [")
         );
@@ -988,7 +982,7 @@ public class MetadataCreateIndexServiceTests extends OpenSearchTestCase {
 
         ClusterState updatedClusterState = clusterStateCreateIndex(
             currentClusterState,
-            org.opensearch.common.collect.Set.of(INDEX_READ_ONLY_BLOCK),
+            Set.of(INDEX_READ_ONLY_BLOCK),
             newIndexMetadata,
             rerouteRoutingTable,
             null
@@ -1027,7 +1021,7 @@ public class MetadataCreateIndexServiceTests extends OpenSearchTestCase {
 
         ClusterState updatedClusterState = clusterStateCreateIndex(
             currentClusterState,
-            org.opensearch.common.collect.Set.of(INDEX_READ_ONLY_BLOCK),
+            Set.of(INDEX_READ_ONLY_BLOCK),
             newIndexMetadata,
             (clusterState, y) -> clusterState,
             metadataTransformer