+++ server/src/test/java/org/opensearch/search/profile/ProfileResultTests.java
@@ -121,26 +121,8 @@ public class ProfileResultTests extends OpenSearchTestCase {
 
     public void testToXContent() throws IOException {
         List<ProfileResult> children = new ArrayList<>();
-        children.add(
-            new ProfileResult(
-                "child1",
-                "desc1",
-                org.opensearch.common.collect.Map.of("key1", 100L),
-                org.opensearch.common.collect.Map.of(),
-                100L,
-                org.opensearch.common.collect.List.of()
-            )
-        );
-        children.add(
-            new ProfileResult(
-                "child2",
-                "desc2",
-                org.opensearch.common.collect.Map.of("key1", 123356L),
-                org.opensearch.common.collect.Map.of(),
-                123356L,
-                org.opensearch.common.collect.List.of()
-            )
-        );
+        children.add(new ProfileResult("child1", "desc1", Map.of("key1", 100L), Map.of(), 100L, List.of()));
+        children.add(new ProfileResult("child2", "desc2", Map.of("key1", 123356L), Map.of(), 123356L, List.of()));
         Map<String, Long> breakdown = new LinkedHashMap<>();
         breakdown.put("key1", 123456L);
         breakdown.put("stuff", 10000L);
@@ -225,14 +207,7 @@ public class ProfileResultTests extends OpenSearchTestCase {
             Strings.toString(builder)
         );
 
-        result = new ProfileResult(
-            "profileName",
-            "some description",
-            org.opensearch.common.collect.Map.of("key1", 12345678L),
-            org.opensearch.common.collect.Map.of(),
-            12345678L,
-            org.opensearch.common.collect.List.of()
-        );
+        result = new ProfileResult("profileName", "some description", Map.of("key1", 12345678L), Map.of(), 12345678L, List.of());
         builder = XContentFactory.jsonBuilder().prettyPrint().humanReadable(true);
         result.toXContent(builder, ToXContent.EMPTY_PARAMS);
         assertEquals(
@@ -248,14 +223,7 @@ public class ProfileResultTests extends OpenSearchTestCase {
             Strings.toString(builder)
         );
 
-        result = new ProfileResult(
-            "profileName",
-            "some description",
-            org.opensearch.common.collect.Map.of("key1", 1234567890L),
-            org.opensearch.common.collect.Map.of(),
-            1234567890L,
-            org.opensearch.common.collect.List.of()
-        );
+        result = new ProfileResult("profileName", "some description", Map.of("key1", 1234567890L), Map.of(), 1234567890L, List.of());
         builder = XContentFactory.jsonBuilder().prettyPrint().humanReadable(true);
         result.toXContent(builder, ToXContent.EMPTY_PARAMS);
         assertEquals(